finish find page

+++ bookkeeping/Classes/Modules/Find (发现)/Bill/View/BillHeader.h
@@ -9,8 +9,8 @@ NS_ASSUME_NONNULL_BEGIN
 
 @interface BillHeader : BaseView
 
-@property (nonatomic, assign) CGFloat income;
-@property (nonatomic, assign) CGFloat pay;
+@property (nonatomic, assign) NSInteger income;
+@property (nonatomic, assign) NSInteger pay;
 
 @end
